requirements.txt
openai>=1.3.9
python-dotenv>=1.0.0
requests>=2.32.4
faiss-cpu>=1.11.0